backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==24.3.0
blinker==1.7.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
deprecation==2.1.0
docutils==0.21.2
exceptiongroup==1.2.2
Flask==3.0.2
flask-cors==5.0.1
flit==3.12.0
flit_core==3.12.0
fonttools==4.40.0
frozenlist==1.5.0
gotrue==2.12.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
multidict==6.4.2
mysql-connector-python==8.0.31
numpy==1.25.0
outcome==1.3.0.post0
packaging==23.1
pandas==2.2.3
Pillow==9.5.0
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
protobuf==3.20.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.11.3
pydantic_core==2.33.1
pyfiglet==0.8.post1
Pygments==2.15.1
PyJWT==2.10.1
pyparsing==3.1.0
PySocks==1.7.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2025.1
realtime==2.4.2
requests==2.32.3
scipy==1.15.1
selenium==4.28.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
termtables==0.2.4
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.65.0
trio==0.28.0
trio-websocket==0.11.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vboxapi==1.0
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.0.1
wsproto==1.2.0
xgboost==2.1.4
yarl==1.19.0
